Tweaks to SAMC code
git-svn-id: https://phycas.svn.sourceforge.net/svnroot/phycas/trunk@1226 b774380d-c90d-0410-997c-aec342e5d6fd

+++ phycas/src/tree_likelihood.cpp
@@ -2861,8 +2861,8 @@ void TreeLikelihood::debugCompressedDataInfo(
 		std::copy(states.begin() + 1, states.end(), std::ostream_iterator<int>(outf," "));
 		outf << "\t";
 
-		unsigned cnt = pattern_counts[i];
-		outf << boost::str(boost::format("%12d\t") % cnt);
+		double cnt = pattern_counts[i]; 
+		outf << boost::str(boost::format("%12g\t") % cnt);
 
 		std::copy(sites.begin(), sites.end(), std::ostream_iterator<unsigned>(outf," "));
 		outf << "\t";
